requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.30
ultralytics>=8.3
pillow>=10.0
httpx>=0.27
python-multipart>=0.0.9
